Trenger ikke lenger å sende over lister av brukere til kafka-kø for minside
